refactor(io): move the last three element handlers off the loader (spec 24 task 3)
